+++ CHANGELOG.md
@@ -6,8 +6,7 @@ All notable changes to the Vezgo SDK across versions will be documented in this 
 
 
 ## [1.1.2]
-- Add support for `multiWallet` flag which allows to connect multiple wallets in one take and return list of connected account and list of errors
-- Update `onConnection(account, message)` to accept additional `message` param, when `multiWallet` flag is set then account and message get multiple accounts and messages splitted by semicolon
+- Add support for `multiWallet` flag which allows to connect multiple wallets in one take and return list of connected `accounts`, `messages` and `wallets`
 
 ## [1.1.1]
 - Added support for fetching orders